+++ src/Fixer/CommentedOutFunctionFixer.php
@@ -22,9 +22,9 @@ use PhpCsFixer\FixerDefinition\FixerDefinition;
 use PhpCsFixer\FixerDefinition\FixerDefinitionInterface;
 use PhpCsFixer\Preg;
 use PhpCsFixer\Tokenizer\Analyzer\FunctionsAnalyzer;
-use PhpCsFixer\Tokenizer\Analyzer\SwitchAnalyzer;
 use PhpCsFixer\Tokenizer\Token;
 use PhpCsFixer\Tokenizer\Tokens;
+use PhpCsFixerCustomFixers\Analyzer\SwitchAnalyzer;
 
 final class CommentedOutFunctionFixer extends AbstractFixer implements ConfigurableFixerInterface
 {