+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v3.1.0
 - Add PhpdocArrayStyleFixer
+- Restore PhpCsFixerCustomFixers\Analyzer\SwitchAnalyzer (as PhpCsFixer\Tokenizer\Analyzer\SwitchAnalyzer got removed in PHP CS Fixer 3.2.0)
 
 ## v3.0.0
 - Drop support for PHP CS Fixer v2